Move SlowCodeStream to private tools
